Suporte ao PIX e outras melhorias
