+++ modules/gateways/paghiper/tmp/billets/index.php
@@ -1,0 +1,2 @@
+<?php
+// Silence is golden.
